moved fel-related functions from lcls environment into new fel submodule
