+++ environments/lcls/__init__.py
@@ -5,6 +5,7 @@ from badger import environment
 from badger.stats import percent_80
 from badger.errors import BadgerEnvObsError, BadgerInterfaceChannelError
 import logging
+from .fel import *
 
 
 class Environment(environment.Environment):
@@ -162,87 +163,6 @@ class Environment(environment.Environment):
             if time_elapsed > self.check_var_timeout:
                 break
 
-    def get_intensity_n_loss(self):
-        # At lcls the repetition is 120 Hz and the readout buf size is 2800.
-        # The last 120 entries correspond to pulse energies over past 1 second.
-        hxr = self.hxr
-
-        points = self.points
-        logging.info(f'Get value of {points} points')
-
-        # Sleep for a while to get enough data
-        try:
-            rate = self.interface.get_value('EVNT:SYS0:1:LCLSBEAMRATE')
-            logging.info(f'Beam rate: {rate}')
-            nap_time = points / (rate * 1.0)
-        except Exception as e:
-            nap_time = 1
-            logging.warn(
-                'Something went wrong with the beam rate calculation. Let\'s sleep 1 second.')
-            logging.warn(f'Exception was: {e}')
-
-        time.sleep(nap_time)
-
-        if hxr:
-            PV_gas = f'GDET:FEE1:{self.fel_channel}:ENRCHSTCUHBR'
-        else:  # SXR
-            PV_gas = 'EM1K0:GMD:HPS:milliJoulesPerPulseHSTCUSBR'
-        PV_loss = self.loss_pv
-        try:
-            results_dict = self.interface.get_values([PV_gas, PV_loss])
-            intensity_raw = results_dict[PV_gas][-points:]
-            loss_raw = results_dict[PV_loss][-points:]
-            ind_valid = ~np.logical_or(np.isnan(intensity_raw), np.isnan(loss_raw))
-            intensity_valid = intensity_raw[ind_valid]
-            loss_valid = loss_raw[ind_valid]
-
-            gas_p80 = percent_80(intensity_valid)
-            gas_mean = np.mean(intensity_valid)
-            gas_median = np.median(intensity_valid)
-            gas_std = np.std(intensity_valid)
-
-            loss_p80 = percent_80(loss_valid)
-
-            return gas_p80, gas_mean, gas_median, gas_std, loss_p80
-        except Exception:  # if average fails use the scalar input
-            if hxr:  # we don't have scalar input for HXR
-                raise BadgerEnvObsError
-            else:
-                gas = self.interface.get_value('EM1K0:GMD:HPS:milliJoulesPerPulse')
-
-                return gas, gas, gas, 0, 0
-
-    def get_loss(self):  # if only loss is observed
-        points = self.points
-        logging.info(f'Get value of {points} points')
-
-        try:
-            rate = self.interface.get_value('EVNT:SYS0:1:LCLSBEAMRATE')
-            logging.info(f'Beam rate: {rate}')
-            nap_time = points / (rate * 1.0)
-        except Exception as e:
-            nap_time = 1
-            logging.warn(
-                'Something went wrong with the beam rate calculation. Let\'s sleep 1 second.')
-            logging.warn(f'Exception was: {e}')
-
-        time.sleep(nap_time)
-
-        PV_loss = self.loss_pv
-        try:
-            loss_raw = self.interface.get_value(PV_loss)[-points:]
-            ind_valid = ~np.isnan(loss_raw)
-            loss_valid = loss_raw[ind_valid]
-            loss_p80 = percent_80(loss_valid)
-
-            return loss_p80
-        except Exception:  # we don't have scalar input for loss
-            raise BadgerEnvObsError
-
-    def is_pulse_intensity_observed(self, observable_names):
-        return len([name for name in observable_names if
-                    name.startswith('pulse_intensity')])
-
     def is_beam_loss_observed(self, observable_names):
         return 'beam_loss' in observable_names
 
@@ -257,9 +177,9 @@ class Environment(environment.Environment):
 
         if observe_gas:
             intensity_p80, intensity_mean, intensity_median, intensity_std, \
-                loss_p80 = self.get_intensity_n_loss()
+                loss_p80 = fel.observables.get_intensity_n_loss(hxr = self.hxr, points = self.points, loss_pv = self.loss_pv, fel_channel = self.fel_channel, interface = self.interface)
         elif observe_loss:
-            loss_p80 = self.get_loss()
+            loss_p80 = fel.observables.get_loss(points = self.points, loss_pv = self.loss_pv, interface = self.interface)
 
         observable_outputs = {}
         mid = self.beamsize_monitor
