+++ environments/lcls/fel/__init__.py
@@ -1,0 +1,1 @@
+#What should go here?
