+++ environments/lcls/fel/configs.yaml
@@ -1,0 +1,8 @@
+---
+name: fel
+version: "0.1"
+dependencies:
+  - numpy
+  - badger-opt
+interface:
+  - epics_raw
